Fixing bottom navigation state issues. The navigation would not pop back to the main stack and then the selected item would be shown incorrectly
