+++ app/src/main/java/de/seemoo/at_tracking_detection/ui/MainActivity.kt
@@ -49,7 +49,7 @@ class MainActivity : AppCompatActivity() {
         // Passing each menu ID as a set of Ids because each
         // menu should be considered as top level destinations.
         val appBarItems: Set<Int> = setOf(
-            R.id.navigation_devices,
+            R.id.navigation_ignoredDevices,
             R.id.navigation_dashboard,
             R.id.navigation_settings
         )
