+++ app/src/main/res/menu/bottom_nav_menu.xml
@@ -9,7 +9,7 @@
         android:title="@string/title_dashboard" />
 
     <item
-        android:id="@+id/navigation_devices"
+        android:id="@+id/navigation_ignoredDevicesFragment"
         android:contentDescription="@string/title_ignored_devices"
         android:icon="@drawable/ic_baseline_list_24"
         android:title="@string/title_ignored_devices" />
