+++ app/src/main/res/navigation/main_navigation.xml
@@ -11,33 +11,25 @@
         android:label="@string/title_dashboard"
         tools:layout="@layout/fragment_dashboard_risk">
         <action
-            android:id="@+id/dashboardToDevicesFragment"
-            app:destination="@id/navigation_devices"
-            app:enterAnim="@anim/slide_in_right"
-            app:exitAnim="@anim/nav_default_exit_anim"
-            app:popEnterAnim="@anim/nav_default_pop_enter_anim"
-            app:popExitAnim="@anim/nav_default_exit_anim" />
-        <action
-            android:id="@+id/dashboardToDeviceMap"
-            app:destination="@id/deviceMapFragment"
-            app:enterAnim="@anim/slide_in_right"
-            app:exitAnim="@anim/nav_default_exit_anim"
-            app:popEnterAnim="@anim/nav_default_pop_enter_anim"
-            app:popExitAnim="@anim/nav_default_exit_anim" />
-        <action
             android:id="@+id/dashboardToScanFragment"
             app:destination="@id/scanFragment"
             app:enterAnim="@anim/slide_in_bottom"
             app:exitAnim="@anim/nav_default_exit_anim"
             app:popEnterAnim="@anim/nav_default_pop_enter_anim"
-            app:popExitAnim="@anim/nav_default_pop_exit_anim" />
+            app:popExitAnim="@anim/nav_default_pop_exit_anim"
+            app:popUpTo="@id/navigation_dashboard"
+            app:popUpToSaveState="true"
+            />
         <action
             android:id="@+id/action_navigation_dashboard_to_riskDetailFragment"
             app:destination="@id/riskDetailFragment"
             app:enterAnim="@anim/slide_in_right"
             app:exitAnim="@anim/nav_default_exit_anim"
             app:popEnterAnim="@anim/nav_default_pop_enter_anim"
-            app:popExitAnim="@anim/nav_default_exit_anim" />
+            app:popExitAnim="@anim/nav_default_exit_anim"
+            app:popUpTo="@id/navigation_dashboard"
+            app:popUpToSaveState="true"
+            />
     </fragment>
 
     <fragment
@@ -47,7 +39,10 @@
         tools:layout="@xml/fragment_settings">
         <action
             android:id="@+id/action_settings_to_about_libs"
-            app:destination="@id/about_libraries" />
+            app:destination="@id/about_libraries"
+            app:popUpTo="@id/navigation_settings"
+            app:popUpToSaveState="true"
+            />
     </fragment>
 
     <include app:graph="@navigation/aboutlibs_navigation" />
@@ -59,14 +54,16 @@
         tools:layout="@layout/fragment_debug">
         <action
             android:id="@+id/action_navigation_debug_to_debugLogFragment"
-            app:destination="@id/debugLogFragment" />
+            app:destination="@id/debugLogFragment"
+            app:popUpToSaveState="false"
+            app:restoreState="true"
+            />
     </fragment>
 
     <fragment
-        android:id="@+id/navigation_devices"
-        android:name="de.seemoo.at_tracking_detection.ui.devices.DevicesFragment"
-        android:label="@string/title_devices"
-        tools:layout="@layout/fragment_devices">
+        android:id="@+id/devices_found"
+        android:name="de.seemoo.at_tracking_detection.ui.devices.FoundDevicesFragment"
+        android:label="@string/title_devices">
         <argument
             android:name="showDevicesFound"
             android:defaultValue="false"
@@ -77,7 +74,13 @@
             app:argType="boolean" />
         <action
             android:id="@+id/action_navigation_devices_to_trackingFragment"
-            app:destination="@id/trackingFragment" />
+            app:destination="@id/trackingFragment"
+            app:enterAnim="@anim/slide_in_right"
+            app:exitAnim="@anim/nav_default_exit_anim"
+            app:popEnterAnim="@anim/nav_default_pop_enter_anim"
+            app:popExitAnim="@anim/nav_default_pop_exit_anim"
+            app:popUpToSaveState="true"
+            />
     </fragment>
 
     <fragment
@@ -129,9 +132,10 @@
             app:popExitAnim="@anim/nav_default_pop_exit_anim" />
         <action
             android:id="@+id/action_riskDetailFragment_to_navigation_devices"
-            app:destination="@id/navigation_devices"
+            app:destination="@id/devices_found"
             app:enterAnim="@anim/slide_in_right"
             app:exitAnim="@anim/nav_default_exit_anim"
+            app:launchSingleTop="false"
             app:popEnterAnim="@anim/nav_default_pop_enter_anim"
             app:popExitAnim="@anim/nav_default_pop_exit_anim">
             <argument
@@ -145,5 +149,24 @@
         android:name="de.seemoo.at_tracking_detection.ui.debug.DebugLogFragment"
         android:label="Debug Logs"
         tools:layout="@layout/fragment_debug_log" />
+    <fragment
+        android:id="@+id/navigation_ignoredDevices"
+        android:name="de.seemoo.at_tracking_detection.ui.devices.DevicesFragment"
+        android:label="DevicesFragment" />
+    <fragment
+        android:id="@+id/navigation_ignoredDevicesFragment"
+        android:name="de.seemoo.at_tracking_detection.ui.devices.IgnoredDevicesFragment"
+        android:label="IgnoredDevicesFragment" >
+        <action
+            android:id="@+id/action_navigation_ignoredDevicesFragment_to_trackingFragment"
+            app:destination="@id/trackingFragment"
+            app:enterAnim="@anim/slide_in_right"
+            app:exitAnim="@anim/nav_default_exit_anim"
+            app:popEnterAnim="@anim/nav_default_pop_enter_anim"
+            app:popExitAnim="@anim/nav_default_pop_exit_anim"
+            app:popUpToSaveState="true"
+            app:popUpTo="@id/navigation_ignoredDevicesFragment"
+            />
+    </fragment>
 
 </navigation>
